Convert Avro emitter to a single IrExtension and use it in the example (IR mode)
Replace the Java/Kotlin Avro LanguageEmitters with one AvroExtension
(IrExtension) that appends an <Type>Avro schema/converter file per record
and enum. It detects the target language from the IR and renders Java or
Kotlin source accordingly, reusing the existing emitter rendering helpers.

Wire the maven-spring-avro example to run the Java IR emitter in IR mode
with the extension (languages/ir/extensionClasses), and update the example
service/test to the new Sender channel API. Update the Avro integration and
plugin-architecture docs to match.
