+++ examples/maven-spring-avro/src/main/java/community/flock/wirespec/examples/maven/avro/AvroExampleService.java
@@ -1,6 +1,7 @@
 package community.flock.wirespec.examples.maven.avro;
 
 import com.eventloopsoftware.kafka.channel.TestAvroRecord;
+import java.util.function.Consumer;
 import org.apache.avro.generic.GenericData;
 import org.springframework.kafka.core.*;
 import org.springframework.kafka.listener.ContainerProperties;
@@ -9,7 +10,7 @@ import org.springframework.kafka.listener.MessageListener;
 import org.springframework.stereotype.Service;
 
 @Service
-public class AvroExampleService implements TestAvroRecord {
+public class AvroExampleService implements TestAvroRecord.Sender {
 
     private static final String TOPIC = "test-avro-record";
 
@@ -25,19 +26,19 @@ public class AvroExampleService implements TestAvroRecord {
     }
 
     @Override
-    public void invoke(com.eventloopsoftware.kafka.model.TestAvroRecord message) {
+    public void testAvroRecord(com.eventloopsoftware.kafka.model.TestAvroRecord message) {
         var template = new KafkaTemplate<>(kafkaProducerFactory);
         var avro = com.eventloopsoftware.kafka.avro.TestAvroRecordAvro.to(message);
         template.send(TOPIC, avro);
     }
 
-    public void listen(String groupId, TestAvroRecord listener) {
+    public void listen(String groupId, Consumer<com.eventloopsoftware.kafka.model.TestAvroRecord> listener) {
         var containerProps = new ContainerProperties(TOPIC);
         containerProps.setGroupId(groupId);
         var container = new KafkaMessageListenerContainer<>(kafkaConsumerFactory, containerProps);
         container.setupMessageListener((MessageListener<String, GenericData.Record>) data -> {
             var message = com.eventloopsoftware.kafka.avro.TestAvroRecordAvro.from(data.value());
-            listener.invoke(message);
+            listener.accept(message);
         });
         container.start();
     }
