+++ src/integration/avro/build.gradle.kts
@@ -59,6 +59,9 @@ kotlin {
         jvmTest {
             dependencies {
                 implementation(project(":src:compiler:core"))
+                implementation(project(":src:compiler:ir"))
+                implementation(project(":src:compiler:emitters:java"))
+                implementation(project(":src:compiler:emitters:kotlin"))
                 implementation(project(":src:integration:wirespec"))
                 implementation(libs.kafka.avro)
                 implementation(libs.spring.boot.test)
