+++ src/site/docs/docs/architecture/architecture-plugins.md
@@ -64,7 +64,9 @@ Gradle takes the class object directly:
 emitterClass.set(MyCustomEmitter::class.java)
 ```
 
-The custom emitter runs *in addition to* any languages selected for the same task — its `Emitted` output is concatenated with the built-ins'. Wirespec's own first-party integrations (`AvroJavaEmitter`, `AvroKotlinEmitter`, `SpringJavaEmitter`) are themselves wired in through this mechanism, so the [Integration](../integration/integration.mdx) pages double as worked examples. A minimal end-to-end project lives at `examples/maven-spring-custom/`.
+The custom emitter runs *in addition to* any languages selected for the same task — its `Emitted` output is concatenated with the built-ins'. A minimal end-to-end project lives at `examples/maven-spring-custom/`.
+
+Wirespec's own first-party integrations are instead wired in as **IR extensions** (`extensionClasses`): an `IrExtension` reshapes or augments the IR of a built-in IR emitter before code generation. The Avro integration's `AvroExtension`, for example, appends an `<Type>Avro` schema/converter class for every model when registered on the Java or Kotlin IR emitter (see the [Integration](../integration/integration.mdx) pages, with a worked example at `examples/maven-spring-avro/`).
 
 ## Shared contract
